Edited code to run with configuration file
Removed hard-coded configuration details from cam.py, cal.py and
the rfi_spectrum.py, rfi_time scripts.

The configuration data is saved in a file system_parameters which
is in the src folder.

With the system_parameters file you can set the ip and port of
the ROACH board you are using and easily edit the run time
parameters without re-compiling the code. You can alter the
values in the system_parameters file, however ensure every
variable has a value as there is no error-handling when
parsing the configuration file.

The config file contains the names of the 3 calibration files for
the RATTY system in use. These file names are under the heading
[analogue_frontend]. The files for atten_gain_map, system_bandpass
and antenna_bandpass are in the cal folder must be comma separated
files with 2 columns. The first line must contain the headings which can
be strings and every line below must contain 2 comma seperated numbers.
Although the files are all .csv files, you must put just the filename
without the file type suffix in the config file.

If there is no antenna_bandpass calibration file, write none for
the antenna_bandpass filename.